requirements.txt
langchain-nvidia-ai-endpoints==0.2.0
jupyter==1.1.0
notebook==7.2.0
openai==1.45.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-cov==5.0.0
ruff==0.6.0
plotly==5.22.0
numpy
anthropic==0.30.0
google-generativeai==0.8.0
streamlit==1.35.0
pandas
scikit-learn
pyyaml==6.0.1

# Performance Evaluation Dependencies
matplotlib

# Node.js dependencies (for package.json)
# openai
# typescript (optional)
